# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Internal APIs for order-service and telegram binding",
        "title": "Stroy1Click Bot API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/telegram/link": {
            "post": {
                "description": "Generates one-time token and returns t.me link with /start \u003ctoken\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram"
                ],
                "summary": "Create Telegram deep-link for binding",
                "parameters": [
                    {
                        "description": "User id to link",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.LinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.LinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/telegram/send": {
            "post": {
                "description": "Finds chat_id by userId and sends formatted order message to user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Send order notification to Telegram user",
                "parameters": [
                    {
                        "description": "Order DTO",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order_orderReceiver.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order_orderReceiver.JSONResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/order_orderReceiver.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/order_orderReceiver.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/order_orderReceiver.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/telegram/updates": {
            "post": {
                "description": "Receives updates from Telegram and binds userId to chatId.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram"
                ],
                "summary": "Telegram webhook endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Secret token",
                        "name": "X-Telegram-Bot-Api-Secret-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Telegram update payload (subset)",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.UpdatesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.JSONResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bot_tgWorker.JSONResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "bot_tgWorker.JSONResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "bot_tgWorker.LinkRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "bot_tgWorker.LinkResponse": {
            "type": "object",
            "properties": {
                "link_url": {
                    "type": "string"
                }
            }
        },
        "bot_tgWorker.TGChat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "bot_tgWorker.TGMessage": {
            "type": "object",
            "properties": {
                "chat": {
                    "$ref": "#/definitions/bot_tgWorker.TGChat"
                },
                "message_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "bot_tgWorker.UpdatesRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/bot_tgWorker.TGMessage"
                },
                "update_id": {
                    "type": "integer"
                }
            }
        },
        "order_orderReceiver.JSONResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "order_orderReceiver.Order": {
            "type": "object",
            "properties": {
                "contactPhone": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "orderItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order_orderReceiver.OrderItem"
                    }
                },
                "orderStatus": {
                    "$ref": "#/definitions/order_orderReceiver.OrderStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "order_orderReceiver.OrderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "order_orderReceiver.OrderStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "Created",
                "Paid",
                "Shipped",
                "Delivered",
                "Canceled"
            ]
        }
    }
}